ollama_version/requirements.txt
phidata
fastapi
yfinance
duckduckgo-search
python-dotenv
uvicorn
openai
groq
packaging
pytest
httpx
python-multipart
